[bugfix] 하단 스타일 틀어짐 교정 #59
